Removed the hcost function and calculateTurnCost function previously added in astar.cpp.
Removed the hcost function and calculateTurnCost function previously added to cbs.hpp.

Modified code previously modified in a_star.hpp to calculate neighbor fscore.

The above strategy (modifying the heuristic function to add a turn cost to the heuristic function for the purpose of reducing the number of turns) is failing.

Translated with DeepL.com (free version)