Fix the logic of state judgment in search function in a_star.hpp.
Since State is a template class, and the State passed in when the cbs algorithm calls a_star to solve for a low-level node, for example, contains not only the x and y attributes, but also the t attribute. Therefore it cannot be assumed that State x == State y means that the two positions are the same, but also contains the condition that they are at the same time t.